Move architecture inside models
